backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.27.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
faster-whisper>=1.1.0
librosa>=0.10.2
pydub>=0.25.1
nltk>=3.9.1
spacy>=3.8.3
scikit-learn>=1.6.0
python-docx>=1.1.2
reportlab>=4.2.5
soundfile>=0.13.1
faster-whisper>=1.1.0
torch>=2.9.1
sentence-transformers>=5.2.0
transformers>=4.41.0
huggingface-hub>=0.20.0
reportlab>=4.2.5
python-docx>=1.1.2
#networkx>=3.6.1
#jinja2>=3.1.6
#MarkupSafe>=2.0
#typing-extensions>=4.10.0
#transformers>=4.40.0
#onnxruntime>=1.14.0
